Fixed communication between servers
Content taken from  the clipboard will now correctly be passed to the
hooks and to other clipboards.
Post-insert hooks now need to be called manually from clients. This was
necessary because of the synchronous nature of all connections made and
it also gives more control to the clients as to what happens to their
data.